backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Add this specific version
python-dotenv==1.0.0
pillow==10.1.0
opencv-python==4.8.1.78
ultralytics==8.0.200
google-generativeai==0.3.2
aiofiles==23.2.1
python-magic==0.4.27
numpy==1.24.3
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
torch==2.0.1
torchvision==0.15.2

# Test dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2